+++ parallel-based-sorting-algorithms/src/main/java/michaeljosh/parallelBrickSort.java
@@ -5,7 +5,7 @@ import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
 
 public class parallelBrickSort {
-    public static void sortArr(int[] arr) {
+    public static void sortArr(long[] arr) {
         int threadNum = arr.length / 2;
         CyclicBarrier barr = new CyclicBarrier(threadNum);
         Thread[] threads = new Thread[threadNum];
@@ -23,18 +23,18 @@ public class parallelBrickSort {
     }
 
     public static void main(String[] args) {
-        int[] arr = {83, 71, 72, 26, 6, 81, 53, 72, 20, 35, 40, 79, 3, 90, 89, 52, 30};
+        long[] arr = {83, 71, 72, 26, 6, 81, 53, 72, 20, 35, 40, 79, 3, 90, 89, 52, 30};
         sortArr(arr);
         System.out.println(Arrays.toString(arr));
     }
 }
 
 class CompareSwapThread implements Runnable {
-    private int[] arr;
+    private long[] arr;
     private int index;
     private CyclicBarrier barr;
 
-    public CompareSwapThread(int[] arr, int index, CyclicBarrier barr) {
+    public CompareSwapThread(long[] arr, int index, CyclicBarrier barr) {
         this.arr = arr;
         this.index = index;
         this.barr = barr;
@@ -44,7 +44,7 @@ class CompareSwapThread implements Runnable {
     public void run() {
         for (int i = 0; i < arr.length; i++) {
             if (arr[index - 1] > arr[index]) {
-                int t = arr[index - 1];
+                long t = arr[index - 1];
                 arr[index - 1] = arr[index];
                 arr[index] = t;
             }
@@ -54,7 +54,7 @@ class CompareSwapThread implements Runnable {
                 e.printStackTrace();
             }
             if (index + 1 < arr.length && arr[index] > arr[index + 1]) {
-                int t = arr[index];
+                long t = arr[index];
                 arr[index] = arr[index + 1];
                 arr[index + 1] = t;
             }
