+++ parallel-based-sorting-algorithms/src/test/java/michaeljosh/parallelBrickSortTest.java
@@ -9,23 +9,23 @@ public class parallelBrickSortTest {
 
     @Test
     public void test1() {
-        int a[] = {3, 7, 4, 8, 6, 2, 1, 5};
+        long a[] = {3, 7, 4, 8, 6, 2, 1, 5};
         sortArr(a);
-        assertArrayEquals(new int[]{1, 2, 3, 4, 5, 6, 7, 8}, a);
+        assertArrayEquals(new long[]{1, 2, 3, 4, 5, 6, 7, 8}, a);
     }
 
     @Test
     public void test2() {
-        int a[] = {5, 10, 15, 25, 20, 3, 2, 1};
+        long a[] = {5, 10, 15, 25, 20, 3, 2, 1};
         sortArr(a);
-        assertArrayEquals(new int[]{1, 2, 3, 5, 10, 15, 20, 25}, a);
+        assertArrayEquals(new long[]{1, 2, 3, 5, 10, 15, 20, 25}, a);
     }
 
     @Test
     public void test3() {
-        int a[] = {3, 7, 9, 8, 6, 2, 1, 5};
+        long a[] = {3, 7, 9, 8, 6, 2, 1, 5};
         sortArr(a);
-        assertArrayEquals(new int[]{1, 2, 3, 5, 6, 7, 8, 9}, a);
+        assertArrayEquals(new long[]{1, 2, 3, 5, 6, 7, 8, 9}, a);
     }
 
 }
